requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
sqlmodel
asyncpg
greenlet
pyjwt
passlib[bcrypt]
email-validator
yfinance